+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 
 - Fixed: Logger `.warn` not working properly.
 - Updated: Node from v16 to v18.
+- Updated: Rework of Router's add/remove command procedures.
+- Updated: Router's Discord command route add and removal within Router's route manage procedures.
+- Updated: Startup's console output to include current node version.
 
 ---